dev-cmd/irb: use irb binary instead of gem
