+++ Library/Homebrew/brew_irbrc
@@ -2,7 +2,9 @@
 #       other IRB config files like `.irbrc`.
 # NOTE: This doesn't work with system Ruby for some reason.
 
-require 'irb/completion'
+require_relative "global"
+require_relative "brew_irb_helpers"
+require "irb/completion"
 
 IRB.conf[:HISTORY_FILE] = "#{Dir.home}/.brew_irb_history"
 IRB.conf[:IRB_NAME] = "brew"
