+++ Library/Homebrew/dev-cmd/irb.rb
@@ -2,37 +2,6 @@
 # frozen_string_literal: true
 
 require "abstract_command"
-require "formula"
-require "formulary"
-require "cask/cask_loader"
-
-class String
-  # @!visibility private
-  sig { params(args: Integer).returns(Formula) }
-  def f(*args)
-    Formulary.factory(self, *args)
-  end
-
-  # @!visibility private
-  sig { params(config: T.nilable(T::Hash[Symbol, T.untyped])).returns(Cask::Cask) }
-  def c(config: nil)
-    Cask::CaskLoader.load(self, config:)
-  end
-end
-
-class Symbol
-  # @!visibility private
-  sig { params(args: Integer).returns(Formula) }
-  def f(*args)
-    to_s.f(*args)
-  end
-
-  # @!visibility private
-  sig { params(config: T.nilable(T::Hash[Symbol, T.untyped])).returns(Cask::Cask) }
-  def c(config: nil)
-    to_s.c(config:)
-  end
-end
 
 module Homebrew
   module DevCmd
@@ -54,8 +23,6 @@ module Homebrew
 
       sig { override.void }
       def run
-        clean_argv
-
         if args.examples?
           puts <<~EOS
             'v8'.f # => instance of the v8 formula
@@ -72,8 +39,6 @@ module Homebrew
         if args.pry?
           Homebrew.install_bundler_gems!(groups: ["pry"])
           require "pry"
-        else
-          require "irb"
         end
 
         require "keg"
@@ -85,24 +50,16 @@ module Homebrew
           Pry.config.history_file = "#{Dir.home}/.brew_pry_history"
           Pry.config.prompt_name = "brew"
 
+          require "brew_irb_helpers"
+
           Pry.start
         else
           ENV["IRBRC"] = (HOMEBREW_LIBRARY_PATH/"brew_irbrc").to_s
 
-          IRB.start
+          $stdout.flush
+          $stderr.flush
+          exec File.join(RbConfig::CONFIG["bindir"], "irb"), "-I", $LOAD_PATH.join(File::PATH_SEPARATOR), *args.named
         end
-      end
-
-      private
-
-      # Remove the `--debug`, `--verbose` and `--quiet` options which cause problems
-      # for IRB and have already been parsed by the CLI::Parser.
-      sig { returns(T.nilable(T::Array[Symbol])) }
-      def clean_argv
-        global_options = Homebrew::CLI::Parser
-                         .global_options
-                         .flat_map { |options| options[0..1] }
-        ARGV.reject! { |arg| global_options.include?(arg) }
       end
     end
   end
